third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-profiler-profiling-expected.txt

Tests that sampling heap profiling works.

Profiler was enabled.

Running: testProfiling
found pageFunction heap-profiler-profiling.js:16
found (anonymous) heap-profiler-profiling.js:1

Profiler was disabled.
